fix: thymeleaf - form 을 위한 프로젝트 수정
