+++ src/main/java/hello/item_service/TestDataInit.java
@@ -1,0 +1,24 @@
+package hello.item_service;
+
+import hello.item_service.domain.item.Item;
+import hello.item_service.domain.item.ItemRepository;
+import jakarta.annotation.PostConstruct;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+
+@Component
+@RequiredArgsConstructor
+public class TestDataInit {
+
+    private final ItemRepository itemRepository;
+
+    /**
+     * 테스트용 데이터 추가
+     */
+    @PostConstruct
+    public void init() {
+        itemRepository.save(new Item("itemA", 10000, 10));
+        itemRepository.save(new Item("itemB", 20000, 20));
+    }
+
+}
