+++ src/main/java/hello/item_service/domain/item/Item.java
@@ -1,20 +1,18 @@
 package hello.item_service.domain.item;
 
-import lombok.Getter;
-import lombok.Setter;
+import lombok.Data;
 
-@Getter @Setter
+@Data
 public class Item {
 
     private Long id;
     private String itemName;
-    private Integer price; // 갯수가 없다는 표식인 NULL 을 넣기 위해 int 대신 Integer 사용
+    private Integer price;
     private Integer quantity;
 
     public Item() {
     }
 
-    // id 를 제외한 생성자
     public Item(String itemName, Integer price, Integer quantity) {
         this.itemName = itemName;
         this.price = price;
