--- src/main/resources/static/html/editForm.html
@@ -1,50 +1,0 @@
-<!DOCTYPE HTML>
-<html>
-<head>
-    <meta charset="utf-8">
-    <link href="../css/bootstrap.min.css" rel="stylesheet">
-    <style>
-        .container {
-            max-width: 560px;
-        }
-    </style>
-</head>
-<body>
-<div class="container">
-    <div class="py-5 text-center">
-        <h2>상품 수정 폼</h2>
-    </div>
-    <form action="item.html" method="post">
-        <div>
-            <label for="id">상품 ID</label>
-            <input type="text" id="id" name="id" class="form-control" value="1" readonly>
-        </div>
-        <div>
-            <label for="itemName">상품명</label>
-            <input type="text" id="itemName" name="itemName" class="form-
-control" value="상품A">
-        </div>
-        <div>
-            <label for="price">가격</label>
-            <input type="text" id="price" name="price" class="form-control" value="10000">
-        </div>
-        <div>
-            <label for="quantity">수량</label>
-            <input type="text" id="quantity" name="quantity" class="form-
-control" value="10">
-        </div>
-        <hr class="my-4">
-        <div class="row">
-            <div class="col">
-                <button class="w-100 btn btn-primary btn-lg" type="submit">저장</
-                button>
-            </div>
-            <div class="col">
-                <button class="w-100 btn btn-secondary btn-lg"
-                        onclick="location.href='item.html'" type="button">취소</button>
-            </div>
-        </div>
-    </form>
-</div> <!-- /container -->
-</body>
-</html>
