--- src/main/resources/static/html/item.html
@@ -1,48 +1,0 @@
-<!DOCTYPE HTML>
-<html>
-<head>
-    <meta charset="utf-8">
-    <link href="../css/bootstrap.min.css" rel="stylesheet">
-    <style>
-        .container {
-            max-width: 560px;
-        }
-    </style>
-</head>
-<body>
-<div class="container">
-    <div class="py-5 text-center">
-        <h2>상품 상세</h2>
-    </div>
-    <div>
-        <label for="itemId">상품 ID</label>
-        <input type="text" id="itemId" name="itemId" class="form-control" value="1" readonly>
-    </div>
-    <div>
-        <label for="itemName">상품명</label>
-        <input type="text" id="itemName" name="itemName" class="form-control" value="상품A" readonly>
-    </div>
-    <div>
-        <label for="price">가격</label>
-        <input type="text" id="price" name="price" class="form-control" value="10000" readonly>
-    </div>
-    <div>
-        <label for="quantity">수량</label>
-        <input type="text" id="quantity" name="quantity" class="form-control" value="10" readonly>
-    </div>
-    <hr class="my-4">
-    <div class="row">
-        <div class="col">
-            <button class="w-100 btn btn-primary btn-lg"
-                    onclick="location.href='editForm.html'" type="button">상품 수정
-            </button>
-        </div>
-        <div class="col">
-            <button class="w-100 btn btn-secondary btn-lg"
-                    onclick="location.href='items.html'" type="button">목록으로
-            </button>
-        </div>
-    </div>
-</div> <!-- /container -->
-</body>
-</html>
