+++ src/main/resources/static/index.html
@@ -8,7 +8,7 @@
 <ul>
     <li>상품 관리
         <ul>
-            <li><a href="/basic/items">상품 관리 - 기본</a></li>
+            <li><a href="/form/items">상품 관리 - 폼(form)</a></li>
         </ul>
     </li>
 </ul>
