--- src/main/resources/templates/basic/addForm.html
@@ -1,50 +1,0 @@
-<!DOCTYPE HTML>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="utf-8">
-    <link th:href="@{/css/bootstrap.min.css}" rel="stylesheet">
-    <style>
-        .container {
-            max-width: 560px;
-        }
-    </style>
-</head>
-<body>
-<div class="container">
-
-    <div class="py-5 text-center">
-        <h2>상품 등록 폼</h2>
-    </div>
-
-    <h4 class="mb-3">상품 입력</h4>
-
-    <!-- action 의 값이 없다면 현재 URL 에 데이터를 전송한다: 같은 URL 이지만 다른 method 이므로 같이 사용가능 -->
-    <form action="item.html" th:action method="post">
-        <div>
-            <label for="itemName">상품명</label>
-            <input type="text" id="itemName" name="itemName" class="form-control" placeholder="이름을 입력하세요">
-        </div>
-        <div>
-            <label for="price">가격</label>
-            <input type="text" id="price" name="price" class="form-control" placeholder="가격을 입력하세요">
-        </div>
-        <div>
-            <label for="quantity">수량</label>
-            <input type="text" id="quantity" name="quantity" class="form-control" placeholder="수량을 입력하세요">
-        </div>
-        <hr class="my-4">
-        <div class="row">
-            <div class="col">
-                <button class="w-100 btn btn-primary btn-lg" type="submit">상품 등록
-                </button>
-            </div>
-            <div class="col">
-                <button class="w-100 btn btn-secondary btn-lg"
-                        onclick="location.href='items.html'" type="button">취소
-                </button>
-            </div>
-        </div>
-    </form>
-</div> <!-- /container -->
-</body>
-</html>
