+++ src/main/resources/templates/form/addForm.html
@@ -1,8 +1,9 @@
 <!DOCTYPE HTML>
-<html>
+<html xmlns:th="http://www.thymeleaf.org">
 <head>
     <meta charset="utf-8">
-    <link href="../css/bootstrap.min.css" rel="stylesheet">
+    <link th:href="@{/css/bootstrap.min.css}"
+          href="../css/bootstrap.min.css" rel="stylesheet">
     <style>
         .container {
             max-width: 560px;
@@ -10,16 +11,17 @@
     </style>
 </head>
 <body>
+
 <div class="container">
+
     <div class="py-5 text-center">
         <h2>상품 등록 폼</h2>
     </div>
-    <h4 class="mb-3">상품 입력</h4>
-    <form action="item.html" method="post">
+
+    <form action="item.html" th:action method="post">
         <div>
             <label for="itemName">상품명</label>
-            <input type="text" id="itemName" name="itemName" class="form-
-control" placeholder="이름을 입력하세요">
+            <input type="text" id="itemName" name="itemName" class="form-control" placeholder="이름을 입력하세요">
         </div>
         <div>
             <label for="price">가격</label>
@@ -27,23 +29,25 @@ control" placeholder="이름을 입력하세요">
         </div>
         <div>
             <label for="quantity">수량</label>
-            <input type="text" id="quantity" name="quantity" class="form-
-control" placeholder="수량을 입력하세요">
+            <input type="text" id="quantity" name="quantity" class="form-control" placeholder="수량을 입력하세요">
         </div>
+
         <hr class="my-4">
+
         <div class="row">
             <div class="col">
-                <button class="w-100 btn btn-primary btn-lg" type="submit">상품 등
-                    록
-                </button>
+                <button class="w-100 btn btn-primary btn-lg" type="submit">상품 등록</button>
             </div>
             <div class="col">
                 <button class="w-100 btn btn-secondary btn-lg"
-                        onclick="location.href='items.html'" type="button">취소
-                </button>
+                        onclick="location.href='items.html'"
+                        th:onclick="|location.href='@{/form/items}'|"
+                        type="button">취소</button>
             </div>
         </div>
+
     </form>
+
 </div> <!-- /container -->
 </body>
 </html>
