+++ src/main/resources/templates/form/editForm.html
@@ -2,7 +2,8 @@
 <html xmlns:th="http://www.thymeleaf.org">
 <head>
     <meta charset="utf-8">
-    <link th:href="@{/css/bootstrap.min.css}" rel="stylesheet">
+    <link th:href="@{/css/bootstrap.min.css}"
+          href="../css/bootstrap.min.css" rel="stylesheet">
     <style>
         .container {
             max-width: 560px;
@@ -10,6 +11,7 @@
     </style>
 </head>
 <body>
+
 <div class="container">
 
     <div class="py-5 text-center">
@@ -19,22 +21,19 @@
     <form action="item.html" th:action method="post">
         <div>
             <label for="id">상품 ID</label>
-            <input type="text" id="id" name="id" class="form-control" th:value="${item.id}" readonly>
+            <input type="text" id="id" name="id" class="form-control" value="1" th:value="${item.id}" readonly>
         </div>
-
         <div>
             <label for="itemName">상품명</label>
-            <input type="text" id="itemName" name="itemName" class="form-control" th:value="${item.itemName}">
+            <input type="text" id="itemName" name="itemName" class="form-control" value="상품A" th:value="${item.itemName}">
         </div>
-
         <div>
             <label for="price">가격</label>
-            <input type="text" id="price" name="price" class="form-control" th:value="${item.price}">
+            <input type="text" id="price" name="price" class="form-control" value="10000" th:value="${item.price}">
         </div>
-
         <div>
             <label for="quantity">수량</label>
-            <input type="text" id="quantity" name="quantity" class="form-control" th:value="${item.quantity}">
+            <input type="text" id="quantity" name="quantity" class="form-control" value="10" th:value="${item.quantity}">
         </div>
 
         <hr class="my-4">
@@ -43,15 +42,16 @@
             <div class="col">
                 <button class="w-100 btn btn-primary btn-lg" type="submit">저장</button>
             </div>
-
             <div class="col">
                 <button class="w-100 btn btn-secondary btn-lg"
                         onclick="location.href='item.html'"
-                        th:onclick="|location.href='@{/basic/items/{itemId}(itemId=${item.id})}'|"
+                        th:onclick="|location.href='@{/form/items/{itemId}(itemId=${item.id})}'|"
                         type="button">취소</button>
             </div>
         </div>
+
     </form>
+
 </div> <!-- /container -->
 </body>
 </html>
