+++ src/main/resources/templates/form/item.html
@@ -2,7 +2,8 @@
 <html xmlns:th="http://www.thymeleaf.org">
 <head>
     <meta charset="utf-8">
-    <link th:href="@{/css/bootstrap.min.css}" rel="stylesheet">
+    <link th:href="@{/css/bootstrap.min.css}"
+            href="../css/bootstrap.min.css" rel="stylesheet">
     <style>
         .container {
             max-width: 560px;
@@ -10,7 +11,9 @@
     </style>
 </head>
 <body>
+
 <div class="container">
+
     <div class="py-5 text-center">
         <h2>상품 상세</h2>
     </div>
@@ -20,35 +23,38 @@
 
     <div>
         <label for="itemId">상품 ID</label>
-        <input type="text" id="itemId" name="itemId" class="form-control" th:value="${item.id}" readonly>
+        <input type="text" id="itemId" name="itemId" class="form-control" value="1" th:value="${item.id}" readonly>
     </div>
     <div>
         <label for="itemName">상품명</label>
-        <input type="text" id="itemName" name="itemName" class="form-control" th:value="${item.itemName}" readonly>
+        <input type="text" id="itemName" name="itemName" class="form-control" value="상품A" th:value="${item.itemName}" readonly>
     </div>
     <div>
         <label for="price">가격</label>
-        <input type="text" id="price" name="price" class="form-control" th:value="${item.price}" readonly>
+        <input type="text" id="price" name="price" class="form-control" value="10000" th:value="${item.price}" readonly>
     </div>
     <div>
         <label for="quantity">수량</label>
-        <input type="text" id="quantity" name="quantity" class="form-control" th:value="${item.quantity}" readonly>
+        <input type="text" id="quantity" name="quantity" class="form-control" value="10" th:value="${item.quantity}" readonly>
     </div>
+
     <hr class="my-4">
+
     <div class="row">
         <div class="col">
             <button class="w-100 btn btn-primary btn-lg"
                     onclick="location.href='editForm.html'"
-                    th:onclick="|location.href='@{/basic/items/{itemId}/edit(itemId=${item.id})}'|"
+                    th:onclick="|location.href='@{/form/items/{itemId}/edit(itemId=${item.id})}'|"
                     type="button">상품 수정</button>
         </div>
         <div class="col">
             <button class="w-100 btn btn-secondary btn-lg"
                     onclick="location.href='items.html'"
-                    th:onclick="|location.href='@{/basic/items}'|"
+                    th:onclick="|location.href='@{/form/items}'|"
                     type="button">목록으로</button>
         </div>
     </div>
+
 </div> <!-- /container -->
 </body>
 </html>
