+++ src/main/resources/templates/form/items.html
@@ -2,22 +2,25 @@
 <html xmlns:th="http://www.thymeleaf.org">
 <head>
     <meta charset="utf-8">
-    <link th:href="@{/css/bootstrap.min.css}" rel="stylesheet">
+    <link th:href="@{/css/bootstrap.min.css}"
+            href="../css/bootstrap.min.css" rel="stylesheet">
 </head>
 <body>
+
 <div class="container" style="max-width: 600px">
     <div class="py-5 text-center">
         <h2>상품 목록</h2>
     </div>
+
     <div class="row">
         <div class="col">
             <button class="btn btn-primary float-end"
                     onclick="location.href='addForm.html'"
-                    th:onclick="|location.href='@{/basic/items/add}'|"
-                    type="button">상품 등록
-            </button>
+                    th:onclick="|location.href='@{/form/items/add}'|"
+                    type="button">상품 등록</button>
         </div>
     </div>
+
     <hr class="my-4">
     <div>
         <table class="table">
@@ -31,14 +34,16 @@
             </thead>
             <tbody>
             <tr th:each="item : ${items}">
-                <td><a href="item.html" th:href="@{/basic/items/{itemId}(itemId=${item.id})}" th:text="${item.id}">회원id</a></td>
-                <td><a href="item.html" th:href="@{|/basic/items/${item.id}|}" th:text="${item.itemName}">상품명</a></td>
+                <td><a href="item.html" th:href="@{/form/items/{itemId}(itemId=${item.id})}" th:text="${item.id}">회원id</a></td>
+                <td><a href="item.html" th:href="@{|/form/items/${item.id}|}" th:text="${item.itemName}">상품명</a></td>
                 <td th:text="${item.price}">10000</td>
                 <td th:text="${item.quantity}">10</td>
             </tr>
             </tbody>
         </table>
     </div>
+
 </div> <!-- /container -->
+
 </body>
 </html>
